requirements.txt
matplotlib
pyserial